API

Exported types

Exported methods

Base.:∘ — Function

(∘)(op::SymOperation, kv::KVec, checkabc::Bool=true) --> KVec

Computes the action of the SymOperation op $≡ g$ on a KVec kv $≡ k$ using that $g$ acts on k-vectors as $k(G)′ = [g(R)ᵀ]⁻¹k(G)$, with $g$ in an $R$-basis and k in a $G$-basis. Returns a new KVec, that is possibly distinct from its original only by a reciprocal lattice vector (i.e. multiple of integers).

If checkabc = false, the free part of KVec is not transformed (can be useful in situation when kabc is zero, and several transformations are requested).

Base.:∘ — Method

(∘)(op1::T, op2::T, modτ::Bool=true) where T<:SymOperation

Compose two symmetry operations op1={W₁|w₁} and op2={W₂|w₂} using the composition rule (in Seitz notation)

{W₁|w₁}{W₂|w₂} = {W₁*W₂|w₁+W₁*w₂}

for symmetry operations opᵢ = {Wᵢ|wᵢ}. By default, the translation part of the {W₁W₂|w₁+W₁w₂} is reduced to the range [0,1], i.e. computed modulo 1. This can be toggled off (or on) by the Boolean flag modτ (enabled, i.e. true, by default). Returns another SymOperation.

Crystalline.bandreps — Method

bandreps(sgnum::Integer; 
         allpaths::Bool=false, spinful::Bool=false, timereversal::Bool=true)

Returns the BandRepSet for space group sgnum (provided by the Bilbao Crystallographic Server; http://www.cryst.ehu.es/cgi-bin/cryst/programs/bandrep.pl)

Keyword arguments:

allpaths: include a minimal sufficient set (false, default) or all (true) k-vectors.

spinful: single- (false, default) or double-valued (true) irreps, as appropriate for spinless and spinful particles, respectively.

timereversal: assume presence (true, default) or absence (false) of time-reversal symmetry.

Crystalline.basisdim — Method

basisdim(BRS::BandRepSet) --> Int64

Computes the dimension of the (linearly independent parts) of a band representation set. This is $dᵇˢ = dᵃⁱ$ in the notation of Po, Watanabe, & Vishwanath, Nature Commun. 8, 50 (2017). In other words, this is the number of linearly independent basis vectors that span the expansions of a band structure or atomic insulator viewed as symmetry-data.

Crystalline.calcfourier — Method

calcfourier(xyz, flat::AbstractFourierLattice) --> Float64

Compute the real part of the function evaluation of flat at a point xyz (a tuple, SVector, or a vector), i.e. return Re[∑ᵢ cᵢexp(2πi𝐆ᵢ⋅𝐫)] with 𝐆ᵢ denoting a 𝐆-vector in an allowed orbit in flat, and cᵢ an associated coefficient (and with 𝐫 ≡ xyz).

Crystalline.centering — Function

centering(sg::SpaceGroup) --> Char
centering(sgnum::Integer, D::Integer=3) --> Char

Determines the conventional centering type of a given space/plane group sg (alternatively specified by its conventional number sgnum and dimensionality D by comparison with the Hermann-Mauguin notation's first letter.

Possible output values, depending on dimensionality D, are (see ITA Sec. 9.1.4):

D=2 ┌ 'p': no centring (primitive)
    └ 'c': face centered

D=3 ┌ 'P': no centring (primitive)
    ├ 'I': body centred (innenzentriert)
    ├ 'F': all-face centred
    ├ 'A', 'B', 'C': one-face centred, (b,c) or (c,a) or (a,b)
    └ 'R': hexagonal cell rhombohedrally centred

Crystalline.classification — Method

classification(BRS::BandRepSet) --> String

Calculate the symmetry indicator classification of a band representation set, meaning the index-classification inferrable on the basis of symmetry alone.

Technically, the calculation answers a question like "what direct product of $Zₙ$ groups is the the quotient group $Xᵇˢ = {BS}/{AI}$ isomorphic to?".

See e.g. Po, Watanabe, & Vishwanath, Nature Commun. 8, 50 (2017) for more information.

Crystalline.conventionalize — Method

conventionalize(Rs′::DirectBasis, cntr::Char) --> Rs::DirectBasis

Transforms a primitive DirectBasis Rs′ into its conventional equivalent Rs, with the transformation dependent on the centering type cntr (P, I, F, R, A, C, and p, c); for centering P and p, the conventional and primive bases coincide.

Crystalline.crystal — Method

crystal(a,b,c,α,β,γ) --> Rs::DirectBasis{3}

Calculate basis vectors R₁, R₂, R₃ in a 3D Cartesian basis for a right-handed coordinate system with specified basis vector lengths a, b, c (associated with R₁, R₂, R₃, respectively) and specified interaxial angles α=∠(R₂,R₃), β=∠(R₃,R₁), γ=∠(R₁,R₂), with ∠≡angle.

For definiteness, the R₁ basis vector is oriented along the x-axis of the Cartesian coordinate system, and the R₂ axis is placed in the xy-plane.

Crystalline.crystal — Method

crystal(a,b,γ) --> DirectBasis{2}

Calculate basis vectors R₁, R₂ in a 2D Cartesian basis for a right-handed coordinate system with specified basis vector lengths a, b (associated with R₁, R₂, respectively) and specified interaxial angle γ=angle(R₁, R₂).

For definiteness, the R₁ basis vector is oriented along the x-axis of the Cartesian coordinate system.

Crystalline.crystal — Method

crystal(a)  --> DirectBasis{1}

Return a one-dimensional crystal with lattice period a.

Crystalline.crystalsystem — Method

crystalsystem(R::DirectBasis)

Determine the crystal system of a point lattice specified in a conventional DirectBasis using Tables 2.1.2.1, 9.1.7.1, & 9.1.7.2 of the International Tables of Crystallography, Volume 1 (ITA). There are 4 crystal systems in 2D and 7 in 3D (see ITA 2.1.2(iii)):

DIM	SYSTEM	CONDITIONS	FREE PARAMS

1D	linear	none	a

2D	square	a=b & γ=90°	a

rectangular	γ=90°	a,b

hexagonal	a=b & γ=120°	a

oblique	none	a,b,γ

3D	cubic	a=b=c & α=β=γ=90°	a

hexagonal	a=b & α=β=90° & γ=120°	a,c

trigonal	a=b & α=β=90° & γ=120°	a,c (a,α for hR)

tetragonal	a=b & α=β=γ=90°	a,c

orthorhombic	α=β=γ=90°	a,b,c

monoclinic	α=γ=90°	a,b,c,β≥90°

triclinic	none	a,b,c,α,β,γ

The DirectBasis input is assumed to use conventional basis vectors; i.e. not necessarily primitive. For primitive basis vectors, the crystal system can be further reduced into 5 Bravais types in 2D and 14 in 3D.

Crystalline.dim — Method

dim(BR::BandRep) --> Int64

Get the number of bands included in a single BandRep BR; i.e. the "band filling" ν discussed in Po's papers.

Crystalline.directbasis — Method

directbasis(sgnum, D=3;    abclims, αβγlims)
directbasis(sgnum, Val(D); abclims, αβγlims) ---> DirectBasis{D}

Generates a (conventional) DirectBasis for a crystal compatible with the space group number sgnum and dimensionality D. Free parameters in the lattice vectors are chosen randomly, with limits optionally supplied in abclims (lengths) and αβγlims (angles). By convention, the length of the first lattice vector (= a) is set to unity, such that the second and third (= b and c) lattice vectors' lengths are relative to the first.

Limits on the relative uniform distribution of lengths b and c can be specified as 2-tuple kwarg abclims; similarly, limits on the angles α, β, γ can be set via αβγlims (only affects oblique, monoclinic, & triclinic lattices).

Crystalline.find_representation — Function

find_representation(symvals::AbstractVector{Number}, 
                    lgirs::AbstractVector{<:AbstractIrrep},
                    αβγ::Union{AbstractVector{<:Real},Nothing}=nothing,
                    assert_return_T::Type{<:Union{Integer, AbstractFloat}}=Int))

                    --> Vector{assert_return_T}

From a vector (or vector of vectors) of symmetry eigenvalues symvals sampled along all the operations of a group gᵢ, whose irreps are contained in irs (evaluated with optional free parameters αβγ), return the multiplicities of each irrep.

Optionally, the multiciplities' element type can be specified via the assert_return_T argument (performing checked conversion; returns nothing if representation in assert_return_T is impossible). This can be useful if one suspects a particular band to transform like a fraction of an irrep (i.e., the specified symmetry data is incomplete).

Extended help

Effectively, this applies the projection operator P⁽ʲ⁾ of each irrep's character set χ⁽ʲ⁾(gᵢ) (j = 1, ... , Nⁱʳʳ) to the symmetry data sᵢ ≡ symvals:

P⁽ʲ⁾  ≡ (dⱼ/|g|) ∑ᵢ χ⁽ʲ⁾(gᵢ)*gᵢ         [characters χ⁽ʲ⁾(gᵢ), irrep dimension dⱼ]
P⁽ʲ⁾s = (dⱼ/|g|) ∑ᵢ χ⁽ʲ⁾(gᵢ)*sᵢ = nⱼ,   [number of bands that transform like jth irrep]

returning the irrep multiplicities mⱼ ≡ nⱼ/dⱼ.

Crystalline.herring — Method

herring(lgir::LGIrrep, sgops::AbstractVector{SymOperation{D}},
        αβγ::Union{Vector{<:Real},Nothing}=nothing)        --> Tuple{Int, Int}

Computes the Herring criterion for a small irrep lgir::LGIrrep, from

[∑ χ({β|b}²)]/[g₀/M(k)]

over symmetry operations {β,b} that take k → -k. Here g₀ is the order of the point group of the space group and M(k) is the order of the star(k) [both in a primitive basis].

The returned value, [∑ χ({β|b}²)]/[g₀/M(k)], is one of three integers in {1,-1,0} corresponding to {real, pseudoreal, complex} reality. We remind that ISOTROPY's indication of the same reality types i {1,2,3}.

The provided space group operations sgops must be the set reduced by primitive translation vectors; i.e. using spacegroup(...) directly is not allowable in general (since the irreps we reference only include these "reduced" operations). This reduced set of operations can be obtained e.g. from the Γ point irreps of ISOTROPY's dataset, or alternatively, from reduce_ops(spacegroup(...), true).

As a sanity check, a value of αβγ can be provided to check for invariance along a symmetry line/plane/general point in k-space. Obviously, the reality type should invariant to this choice.

Implementation: See e.g. Inui's Eq. (13.48), Dresselhaus, p. 618, and and Herring's original paper at https://doi.org/10.1103/PhysRev.52.361. We mainly followed Cornwell, p. 150-152 & 187-188.

Crystalline.interpolate_kvpath — Method

interpolate_kvpath(kvs::AbstractVector{<:AbstractVector{<:Real}}, Ninterp::Integer) 
    --> Vector{Vector{Float64}}, Int64

Computes an interpolated $k$-path between the discrete $k$-points in kvs, so that the interpolated path has approximately Ninterp points in total (typically fewer).

Since the actual number of points in the path may deviate from the requested Ninterp, the actual number of points in the path is returned along with the interpolated itself.

Note that, in general, it is not possible to do this so that all interpolated $k$-points are equidistant; but points are equidistant in-between the initial discrete points provided in kvs.

See also: splice_kvpath.

Crystalline.isnormal — Method

isnormal(opsᴳ::T, opsᴴ::T; verbose::Bool=false) --> Bool

Determine whether the operations in group H are normal in the group G (each with operations opsᴳ and opsᴴ, respectively, of type T::AbstractVector{SymOperation{D}}), in the sense that

ghg⁻¹ ∈ H ∀ g∈G, h∈H

Returns a Boolean answer (true if normal, false if not).

Note: This that this compares space groups rather than space group types, i.e. the comparison assumes a matching setting choice between H and G. To compare space group types with different conventional settings, they must first be transformed to a shared setting.

Crystalline.israyrep — Function

israyrep(lgir::LGIrrep, αβγ=nothing) -> (::Bool, ::Matrix)

Computes whether a given little group irrep ir is a ray representation by computing the coefficients αᵢⱼ in DᵢDⱼ=αᵢⱼDₖ; if any αᵢⱼ differ from unity, we consider the little group irrep a ray representation (as opposed to the simpler "vector" representations where DᵢDⱼ=Dₖ). The function returns a boolean (true => ray representation) and the coefficient matrix αᵢⱼ.

Crystalline.issubgroup — Method

issubgroup(opsᴳ::T, opsᴴ::T) --> Bool

Determine whether the operations in group H are a subgroup of the group G (each with operations opsᴳ and opsᴴ, respectively, of type T::AbstractVector{SymOperation{D}}), i.e. whether H<G. Specifically, this requires that G and H are both groups and that for every h∈H there exists an element g∈G such that h=g.

Returns a Boolean answer (true if normal, false if not).

Note: This compares space groups rather than space group types, i.e. the comparison assumes a matching setting choice between H and G. To compare space group types with different conventional settings, they must first be transformed to a shared setting.

Crystalline.issymmorph — Function

issymmorph(sgnum::Integer, D::Integer=3) --> Bool

Checks whether a given space group sgnum (of dimensionality D) is symmorphic (true) or nonsymmorphic (false).

Crystalline.issymmorph — Method

issymmorph(op::SymOperation, cntr::Char) --> Bool

Checks whether a given symmetry operation op is symmorphic (true) or nonsymmorphic (false). The operation is assumed to be given in a conventional basis; but the check requires that the translation is zero in a primitive basis. Accordingly, the centering cntr must provided.

Crystalline.issymmorph — Method

issymmorph(sg::AbstractGroup) --> Bool

Checks whether a given space group sg is symmorphic (true) or nonsymmorphic (false).

Crystalline.iuc — Function

iuc(sgnum::Integer, D::Integer=3) --> String

Returns the IUC (International Union of Crystallography) notation for space group number sgnum and dimensionality D, as used in the International Tables of Crystallography. The notation is sometimes also known as the Hermann-Mauguin notation; the functionality is consequently aliased by hermannmauguin(sgnum, D). IUC/Hermann-Mauguin notation applies in one, two, and three-dimensions.

For more information, see https://en.wikipedia.org/wiki/Hermann%E2%80%93Mauguin_notation.

Crystalline.levelsetlattice — Function

levelsetlattice(sgnum::Integer, D::Integer=2, idxmax::NTuple=ntuple(i->2,D))
    --> UnityFourierLattice{D}

Compute a "neutral"/uninitialized Fourier lattice basis, a UnityFourierLattice, consistent with the symmetries of the space group sgnum in dimension D. The resulting lattice flat is expanded in a Fourier basis split into symmetry-derived orbits, with intra-orbit coefficients constrained by the symmetries of the space-group. The inter-orbit coefficients are, however, free and unconstrained.

The Fourier resolution along each reciprocal lattice vector is controlled by idxmax: e.g., if D = 2 and idxmax = (2, 3), the resulting Fourier lattice may contain reciprocal lattice vectors (k₁, k₂) with k₁∈[0,±1,±2] and k₂∈[0,±1,±2,±3], referred to a 𝐆-basis.

This "neutral" lattice can, and usually should, be subsequently modulated by modulate (modulates the inter-orbit coefficients, which will often eliminate symmetries that may remain in the "neutral" configuration, where all inter-orbit coefficients are unity).

Examples

Compute a UnityFourierLattice, modulate it with random inter-orbit coefficients via modulate, and finally plot it (requires using PyPlot):

julia> uflat = levelsetlattice(16, 2)
julia> flat  = modulate(uflat)
julia> Rs    = directbasis(16, 2) 
julia> plot(flat, Rs)

Crystalline.matrix — Function

matrix(BRS::BandRepSet[, includedim::Bool=false])

Return a matrix representation of BRS::BandRepSet, with band representations as columns and irreps over rows.

For includedim=true the band filling (i.e. dim.(BRS)) is included as the last row.

Crystalline.modulate — Method

modulate(flat::UnityFourierLattice{D},
modulation::AbstractVector{ComplexF64}=rand(ComplexF64, length(getcoefs(flat))),
expon::Union{Nothing, Real}=nothing)
                        --> ModulatedFourierLattice{D}

Derive a concrete, modulated Fourier lattice from flat, a UnityFourierLattice struct (that contains the interrelations between orbit coefficients), by multiplying the "normalized" orbit coefficients by a modulation, a complex modulating vector (in general, should be complex; otherwise restores unintended symmetry to the lattice). Distinct modulation vectors produce distinct realizations of the same lattice described by the original flat. By default, a random complex vector is used.

An exponent expon can be provided, which introduces a penalty term to short- wavelength features (i.e. high-|G| orbits) by dividing the orbit coefficients by |G|^expon; producing a more "localized" and "smooth" lattice boundary when expon > 0 (reverse for expon < 0). This basically amounts to a continuous "simplifying" operation on the lattice (it is not necessarily a smoothing operation; it simply suppresses "high-frequency" components). If expon = nothing, no rescaling is performed.

The normscale(!) methods exists to perform subsequent expon norm-rescaling of a ModulatedFourierLattice.

Crystalline.multtable — Method

multtable(ops::AbstractVector{<:SymOperation{D}})

Compute the multiplication (or Cayley) table of a set of symmetry operations. A MultTable is returned, which contains symmetry operations resulting from composition of row ∘ col operators; the table of indices give the symmetry operators relative to the ordering of ops.

Crystalline.normscale! — Method

normscale!(flat::ModulatedFourierLattice, expon::Real) --> ModulatedFourierLattice

In-place equivalent of normscale: changes flat.

Crystalline.normscale — Method

normscale(flat::ModulatedFourierLattice, expon::Real) --> ModulatedFourierLattice

Applies subsequent norm-rescaling via expon; see detailed description in modulate. An in-place variant is provided as normscale!.

Crystalline.pointgroup — Method

pointgroup(ops:AbstractVector{SymOperation{D}})
pointgroup(sg::AbstractGroup)

Computes the point group associated with a space group sg (characterized by a set of operators ops, which, jointly with lattice translations generate the space group), obtained by "taking away" any translational parts and then reducing to the resulting unique rotational operations. (technically, in the language of Bradley & Cracknell, this is the so-called isogonal point group of sg; see Sec. 1.5).

Returns a Vector of SymOperations.

Crystalline.prepare_mpbcalc! — Method

prepare_mpbcalc!(...)

Formats a set of parameters that uniquely specify an MPB calculation, given a space group number sgnum, a Fourier lattice flat, a DirectBasis Rs, a filling fraction filling for flat, interior and exterior (above, below the contour) permittivities εin and εout, as well as a list of k-vectors kvecs, an identifying tag id (to label the calculation for book-keeping purposes), a resolution for the MPB calculation res, and a selection of calculation type runtype ("all", "te", or "tm"). The results are written to requested IO io.

Our preferred choice is to write these parameters to a bash file, with a name generated by the mpb_calcname(...) method.

The options are expected to be fed to the fourier-lattice.ctl file, e.g. through a bash script of the following kind:

    IFS=$'\n'; # stop command-substitutions from word-splitting at space

    PATH_TO_MPB_EXECUTABLE \
        (cat ${calcname}.sh)
        ctl/fourier-lattice.ctl 2>&1 | tee logs/${calcname}.log
        
    unset IFS; # restore usual command-substitution word-splitting practice

on gbar, PATH_TO_MPB_EXECUTABLE is ~/postdoc/mpb-transform-dev/1.8-dev/bin/mpb. Locally, in mpb-ctl we have a file run-fourier-lattice.sh which performs the above, with calcname specified as an input parameter (assumed to be a subfolder /input/).

Crystalline.primitivize — Method

primitivize(flat::AbstractFourierLattice, cntr::Char) --> AbstractFourierLattice

Given flat referred to a conventional basis with centering cntr, compute the derived (but physically equivalent) lattice flat′ referred to the associated primitive basis.

Specifically, if flat refers to a direct conventional basis Rs≡(𝐚 𝐛 𝐜) [with coordinate-vectors 𝐫≡(r₁, r₂, r₃)ᵀ] then flat′ refers to a direct primitive basis Rs′≡(𝐚′ 𝐛′ 𝐜′)≡(𝐚 𝐛 𝐜)P [with coordinate-vectors 𝐫′≡(r₁′, r₂′, r₃′)ᵀ=P⁻¹𝐫], where P denotes the basis-change matrix obtained from primitivebasismatrix(...).

To compute the associated primitive basis vectors, see primitivize(::DirectBasis) [specifically, Rs′ = primitivize(Rs, cntr)].

Examples

A centered ('c') lattice from plane group 5 in 2D, plotted in its conventional and primitive basis (requires using PyPlot):

julia> sgnum = 5; D = 2; cntr = centering(sgnum, D)  # 'c' (body-centered)

julia> Rs   = directbasis(sgnum, D)     # conventional basis (rectangular)
julia> flat = levelsetlattice(sgnum, D) # Fourier lattice in basis of Rs
julia> flat = modulate(flat)            # modulate the lattice coefficients
julia> plot(flat, Rs)

julia> Rs′   = primitivize(Rs, cntr)    # primitive basis (oblique)
julia> flat′ = primitivize(flat, cntr)  # Fourier lattice in basis of Rs′
julia> plot(flat′, Rs′)

Crystalline.primitivize — Method

primitivize(Vs::Basis, sgnum::Integer) --> Rs′::Basis

Transforms a conventional Basis (either DirectBasis or ReciprocalBasis) Vs into its primitive equivalent Vs′, provided that its centering differs from the conventional (P or p), by inferring the Bravais type from the space group number sgnum and applying an applying an appropriate (Basis-type specific) transformation.

Crystalline.primitivize — Method

primitivize(Rs::DirectBasis, cntr::Char) --> Rs′::DirectBasis

Transforms a conventional DirectBasis Rs into its primitive equivalent Rs′, with the transformation dependent on the centering type cntr (P, I, F, R, A, C, and p, c); for centering P and p, the conventional and primive bases coincide.

Crystalline.primitivize — Method

primitivize(Gs::ReciprocalBasis, cntr::Char) --> Gs′::ReciprocalBasis

Calculates the primitive reciprocal basis associated with a ReciprocalBasis Gs derived from a lattice with centering type cntr.

Crystalline.primitivize — Method

primitivize(op::SymOperation, cntr::Char, modw::Bool=true) --> SymOperation

Transforms a symmetry operation op $= {W|w}$ from a conventional cell to a primitive cell (specified by its centering character cntr), then denoted ${W′|w′}$; i.e. performs a basis change ${W′|w′} = {P|p}⁻¹{W|w}{P|p}$ where $P$ and $p$ describe basis change and origin shifts, respectively, of the coordinate transformation.

By default, translation parts of op′, i.e. $w′$ are reduced modulo 1 (modw = true); to disable this, set modw = false.

For additional details, see ITA6 Sec. 1.5.2.3, p. 84.

Crystalline.realify — Method

realify(lgirs::AbstractVector{<:LGIrrep}, verbose::Bool=false)
                                                    --> AbstractVector{<:LGIrrep}

From lgirs, a vector of LGIrreps, determine the associated (gray) co-representations, i.e. the "real", or "physical" irreps that are relevant in scenarios with time-reversal symmetry.

For LGIrrep that are real (type=1), or that characterize a k-point 𝐤 which is not equivalent to -𝐤 (i.e. its star does not include both 𝐤 and -𝐤; equivalently, the little group includes time-reversal symmetry), the associated co-representations are just the original irreps themselves. For pseudo-real (type=2) and complex (type=3) LGIrreps where ±𝐤 are equivalent, the associated co-representations are built from pairs of irreps that "stick" together. This method computes this pairing and sets the LGIrrep field iscorep to true, to indicate that the resulting "paired irrep" (i.e. the co-representation) should be doubled with itself (pseudo-real type) or its complex conjugate (complex type).

Background

For background, see p. 650-652 (and 622-626 for point groups) in Bradley & Cracknell's book. Their discussion is for magnetic groups (the "realified" irreps are really correspond to co-representations of "gray" magnetic groups). Cornwell's book also does a good job of explicating this, as does Inui (p. 296-299).

Keyword arguments

verbose::Bool: if set to true, prints details about mapping from small irrep to small

corep for each LGIrrep (default: false).

Crystalline.reciprocalbasis — Method

reciprocalbasis(Rs::DirectBasis{D}) --> Gs::ReciprocalBasis{D}

Calculates the reciprocal basis associated with a DirectBasis Rs (alternatively supplied as an NTuple of Vectors). Returns a ReciprocalBasis.

Crystalline.schoenflies — Method

schoenflies(sgnum::Integer) --> String

Returns the Schoenflies notation for a given space group number sgnum. Schoenflies notation only applies to point groups and space groups, not plane groups, so this notation is only relevant in three dimensions.

Crystalline.seitz — Method

seitz(op::SymOperation) --> String

Computes the correponding Seitz notation for a symmetry operation in triplet/xyzt form.

Implementation based on ITA5 Table 11.2.1.1, with 3D point group parts inferred from the trace and determinant of the matrix $W$ in the triplet ${W|w}$.

detW\trW	-3	-2	-1	0	1	2	3

1			2	3	4	6	1

-1	-1	-6	-4	-3	m

with the elements of the table giving the type of symmetry operation in in Hermann-Mauguin notation. The rotation axis and the rotation sense are computed following the rules in ITA6 Sec. 1.2.2.4(1)(b-c). The implementation has been checked against the Tables 1.4.2.1-5 of ITA6.

Note that the orientation of axis (i.e. its sign) is not necessarily equal to the orientation picked in those tables; it is a matter of convention, and the conventions have not been explicated in ITA6.

2D operations are treated by the same procedure, by elevation in a third dimension; 1D operations by a simple inspection of sign.

Crystalline.spacegroup — Method

spacegroup(sgnum::Integer, D::Integer=3) --> SpaceGroup{D}

Return the space group symmetry operations in for a given space group number sgnum and dimensionality D as a SpaceGroup{D}. The returned symmetry operations are specified relative to the conventional basis choice, i.e. are not necessarily primitive (see centering). If desired, operations for the primitive unit cell can be subsequently generated using primitivize or reduce_ops.

The default choices for the conventional basis vectors are specified in Bilbao as:

Unique axis b (cell choice 1) for space groups within the monoclinic system.

Obverse triple hexagonal unit cell for R space groups.

Origin choice 2: inversion center at (0,0,0). (Relevant for the centrosymmetric space groups where there are two origin choices, in the orthorhombic, tetragonal and cubic systems)

See also directbasis.

Crystalline.splice_kvpath — Method

splice_kvpath(kvs::AbstractVector{<:AbstractVector{<:Real}}, Nsplice::Integer) 
                                                            --> Vector{Vector{Float64}}

Computes an interpolated $k$-path between the discrete $k$-points in kvs, inserting Nsplice points between each pair of adjacent $k$-points.

See also interpolate_kvpath.

Crystalline.subduction_count — Method

subduction_count(Dᴳᵢ, Dᴴⱼ[, αβγᴴⱼ]) --> Int64

For two groups G and H, where H is a subgroup of G, i.e. G>H, with associated irreducible representations Dᴳᵢ(g) and Dᴴⱼ(h) for elements g∈G and h∈H<G, compute the compatibility relation between the two irreps from the subduction reduction formula (or "magic" formula/Schur orthogonality relation): this is essentially how many times nᴳᴴᵢⱼ the subduced representation Dᴳᵢ↓H contains the irrep Dᴴⱼ; in other words, this gives the compatibility between the two irreps.

Optionally, a vector αβγᴴⱼ may be provided, to evaluate the characters/irreps of Dᴳᵢ at a concrete value of αβγ. This is e.g. meaningful for LGIrreps at non- special k-vectors. Defaults to nothing.

The reduction formula [e.g. Eq. (15) of https://arxiv.org/pdf/1706.09272.pdf] is:

    nᴳᴴᵢⱼ = |H|⁻¹∑₍ₕ₎ χᴳᵢ(h)χᴴⱼ(h)*

As an example, consider space group 207 and the two compatible k-vectors Γ (a point) and Σ (a line):

    lgirsd  = get_lgirreps(207, Val(3));
    Γ_lgirs = lgirsd["Γ"]; # at Γ ≡ [0.0, 0.0, 0.0]
    Σ_lgirs = lgirsd["Σ"]; # at Σ ≡ [α, α, 0.0]

We can test their compatibility like so:

    [[subduction_count(Γi, Σj) for Γi in Γ_lgirs] for Σj in Σ_lgirs]
    > # Γ₁ Γ₂ Γ₃ Γ₄ Γ₅
    >  [ 1, 0, 1, 1, 2] # Σ₁
    >  [ 0, 1, 1, 2, 1] # Σ₂

This entails the following compatibility relations between irreps at Γ and Σ:

    Γ₁ → Σ₁           degeneracies: 1 → 1
    Γ₂ → Σ₂                         1 → 1
    Γ₃ → Σ₁ + Σ₂                    2 → 1 + 1
    Γ₄ → Σ₁ + 2Σ₂                   3 → 1 + 2
    Γ₅ → 2Σ₁ + Σ₂                   3 → 2 + 1

where, in this case, all the small irreps are one-dimensional.

Crystalline.transform — Method

transform(op::SymOperation, P::Matrix{<:Real}, 
          p::Union{Vector{<:Real}, Nothing}=nothing,
          modw::Bool=true)                          --> SymOperation

Transforms a symmetry operation op $= {W|w}$ by a rotation matrix P and a translation vector p (can be nothing for zero-translations), producing a new symmetry operation op′ $= {W′|w′}$: (see ITA6, Sec. 1.5.2.3.)

${W′|w′} = {P|p}⁻¹{W|w}{P|p}$ with $W′ = P⁻¹WP$ and $w′ = P⁻¹(w+Wp-p)$

By default, the translation part of op′, i.e. $w′$, is reduced to the range $[0,1)$, i.e. computed modulo 1. This can be disabled by setting modw = false (default, modw = true).

See also primitivize and conventionalize.

Unexported API

Methods

Crystalline.BandRep — Type

struct BandRep <: AbstractArray{Int64,1}

wyckpos::String

sitesym::String

label::String

dim::Integer

decomposable::Bool

spinful::Bool

irvec::Array{Int64,1}

irlabs::Array{String,1}

Crystalline.BandRepSet — Type

struct BandRepSet <: AbstractArray{BandRep,1}

sgnum::Integer

bandreps::Array{BandRep,1}

kvs::Array{KVec,1}

klabs::Array{String,1}

irlabs::Array{String,1}

allpaths::Bool

spinful::Bool

timeinvar::Bool

Crystalline.CharacterTable — Type

struct CharacterTable{D}

ops::Array{SymOperation{D},1} where D

irlabs::Array{String,1}

chartable::Array{Complex{Float64},2}

tag::String

Crystalline.CharacterTable — Method

CharacterTable(irs::AbstractVector{<:AbstractIrrep}, αβγ=nothing)

Returns a CharacterTable associated with vector of AbstractIrreps irs.

Optionally, an αβγ::AbstractVector{<:Real} variable can be passed to evaluate the irrep (and associated characters) with concrete free parameters (e.g., for LGIrreps, a concrete k-vector sampled from a "line-irrep"). Defaults to nothing, indicating it being either irrelevant (e.g., for PGIrreps) or all free parameters implicitly set to zero.

Crystalline.DirectBasis — Type

DirectBasis{D} <: Basis{D}

vecs:NTuple{D, SVector{D, Float64}}

Crystalline.KVec — Type

struct KVec

k₀::Array{Float64,1}

kabc::Array{Float64,2}

Crystalline.KVec — Method

KVec(str::AbstractString) --> KVec

Construct a KVec struct from a string representations of a k-vector, supplied in either of the formats "($x,$y,$z)", "[$x,$y,$z]", "$x,$y,$z", where the coordinates x,y, and z are strings that can contain fractions, decimal numbers, and "free" parameters {'α','β','γ'} (or, alternatively, {'u','v','w'}). Returns the associated KVec.

Any "fixed"/constant part of a coordinate must precede any free parts, e.g., x="1+α" is allowable but x="α+1" is not.

Fractions such as 1/2 can be parsed: but use of any other special operator besides / will result in faulty operations (e.g. do not use *).

Crystalline.LGIrrep — Type

struct LGIrrep{D} <: Crystalline.AbstractIrrep{D}

cdml::String

lg::LittleGroup{D} where D

matrices::Array{Array{Complex{Float64},2},1}

translations::Array{Array{Float64,1},1}

type::Int64

iscorep::Bool

Crystalline.LittleGroup — Type

struct LittleGroup{D} <: Crystalline.AbstractGroup{D}

num::Int64

kv::KVec

klab::String

operations::Array{SymOperation{D},1} where D

Crystalline.ModulatedFourierLattice — Type

ModulatedFourierLattice{D} <: AbstractFourierLattice{D}

A D-dimensional concrete Fourier/plane wave lattice, derived from a UnityFourierLattice by scaling/modulating its orbit coefficients by complex numbers; in general, the coefficients do not have unit norm.

Crystalline.MultTable — Type

struct MultTable{D} <: AbstractArray{Int64,2}

operations::Array{SymOperation{D},1} where D

indices::Array{Int64,2}

isgroup::Bool

Crystalline.PGIrrep — Type

struct PGIrrep{D} <: Crystalline.AbstractIrrep{D}

cdml::String

pg::PointGroup{D} where D

matrices::Array{Array{Complex{Float64},2},1}

type::Int64

Crystalline.PointGroup — Type

struct PointGroup{D} <: Crystalline.AbstractGroup{D}

num::Int64

label::String

operations::Array{SymOperation{D},1} where D

Crystalline.ReciprocalBasis — Type

ReciprocalBasis{D} <: Basis{D}

vecs:NTuple{D, SVector{D, Float64}}

Crystalline.SpaceGroup — Type

struct SpaceGroup{D} <: Crystalline.AbstractGroup{D}

num::Int64

operations::Array{SymOperation{D},1} where D

Crystalline.SymOperation — Type

struct SymOperation{D} <: AbstractArray{Float64,2}

matrix::Array{Float64,2}

Crystalline.UnityFourierLattice — Type

UnityFourierLatticeFourierLattice{D} <: AbstractFourierLattice{D}

A general D-dimensional Fourier/plane wave lattice (specified by G-orbits and coefficient interrelations); specifies the allowable interrelations between coefficients within each orbit. The norm of all orbit coefficients is unity. The G-orbits orbits (& associated coefficients) are sorted in order of increasing |G| (low to high).

Base.inv — Method

inv(op::SymOperation{D}) --> SymOperation{D}

Compute the inverse {W|w}⁻¹ ≡ {W⁻¹|-W⁻¹w} of an operator op ≡ {W|w}.

Base.isapprox — Method

isapprox(kv1::KVec, kv2::KVec[, cntr::Char]; kwargs...) --> Bool

Compute approximate equality of two KVec's k1 and k2 modulo any primitive G-vectors. To ensure that primitive G-vectors are used, the centering type cntr (see centering(cntr, dim)) must be given (the dimensionality is inferred from kv1 and kv2). Optionally, keyword arguments (e.g., atol and rtol) can be provided, to include in calls to Base.isapprox.

If cntr is not provided, the comparison will not account for equivalence by primitive G-vectors.

Crystalline.:⊚ — Method

(⊚)(op1::T, op2::T) where T<:SymOperation -->  Vector{Float64}

Compose two symmetry operations op1={W₁|w₁} and op2={W₂|w₂} and return the quotient of w₁+W₁*w₂ and 1. This functionality complements op1∘op2, which yields the translation modulo 1; accordingly, translation(op1∘op2) + op1⊚op2 yields the translation component of the composition op1 and op2 without taking it modulo 1, i.e. including any "trivial" lattice translation.

Note that ⊚ can be auto-completed in Julia via \circledcirc+[tab]

Crystalline._findsubgroup — Method

_findsubgroup(opsᴳ, opsᴴ) --> Bool, Vector{Int64}

Determine whether the group H (with operators opsᴴ) is a subgroup of the group G (with operators opsᴳ), i.e. whether H<G, and returns an indexing vector idxs of opsᴳ into opsᴴ (empty if false), such that opsᴳ[idxs] ≡ H. The first return argument is a Boolean (whether H<G); the second is idxs.

Crystalline.basis2matrix — Method

basis2matrix(Vs::Basis{D}) where D

Compute a matrix [Vs[1] Vs[2] .. Vs[D]] from Vs::Basis{D}, i.e. a matrix whose columns are the basis vectors in Vs.

Note: Trying to use the iteration interface via hcat(Vs...) does not lead to a correctly inferred type Matrix::Float64 (and a type-assertion does not improve speed much). Instead, we just use the .vec field of Vs directly, which achieves good performance.

Crystalline.cartesianize — Method

cartesianize(op::SymOperation{D}, Rs::DirectBasis{D}) --> SymOperation{D}

Convert a SymOperation opˡ from the lattice basis to a Cartesian basis, by computing the transformed operators opᶜ = 𝐑*opˡ*𝐑⁻¹ via the Cartesian basis matrix 𝐑 (whose columns are the DirectBasis vectors Rs[i]).

Note 1

The matrix 𝐑 maps vectors coefficients in a lattice basis 𝐯ˡ to coefficients in a Cartesian basis 𝐯ᶜ as 𝐯ˡ = 𝐑⁻¹𝐯ᶜ and vice versa as 𝐯ᶜ = 𝐑𝐯ˡ. Since a general transformation P transforms an "original" vectors with coefficients 𝐯 to new coefficients 𝐯′ via 𝐯′ = P⁻¹𝐯 and since we here here consider the lattice basis as the "original" bais we have P = 𝐑⁻¹. As such, the transformation of the operator op transforms as opᶜ = P⁻¹*opˡ*P, i.e. opᶜ = transform(opˡ,P) = transform(opˡ,𝐑⁻¹).

Note 2

The display (e.g. Seitz and xyzt notation) of SymOperations e.g. in the REPL implicitly assumes integer coefficients for its point-group matrix: as a consequence, displaying SymOperations in a Cartesian basis may yield undefine behavior. The matrix representation remains valid, however.

Crystalline.compact_print_matrix — Function

compact_print_matrix(io, X, prerow)
compact_print_matrix(io, X, prerow, elformat)
compact_print_matrix(io, X, prerow, elformat, sep)

Canibalized and adapted from Base.print_matrix, specifically to allow a prerow input.

Should never be used for printing very large matrices, as it will not wrap or abbreviate rows/columns.

Crystalline.compatibility — Method

compatibility(lgirsd)

TODO: Seems entirely broken? Not sure what this is supposed to do.

Crystalline.connectivity — Method

connectivity(lgirsd)

Crystalline.find_compatible_kvec — Method

find_compatible_kvec(kv, kvs′)

Crystalline.findequiv — Method

findequiv(op::SymOperation, ops::AbstractVector{SymOperation{D}}, cntr::Char) 
                                            --> Tuple{Int, Vector{Float64}}

Search for an operator op′ in ops which is equivalent, modulo differences by primitive lattice translations Δw, to op. Return the index of op′ in ops, as well as the primitive translation difference Δw. If no match is found returns (nothing, nothing).

The small irreps of op at wavevector k, Dⱼᵏ[op], can be computed from the small irreps of op′, Dⱼᵏ[op′], via Dⱼᵏ[op] = exp(2πik⋅Δw)Dⱼᵏ[op′]

Crystalline.fractionify! — Function

fractionify!(io::IO, x::Real, forcesign::Bool=true, tol::Real=1e-4)

Write a string representation of the nearest fraction (within a tolerance tol) of x to io. If forcesign is true, the sign character of x is printed whether + or - (otherwise, only printed if -).

Crystalline.is_compatible_kvec — Method

is_compatible_kvec(kv, kv′)

Check whether a special k-point kv is compatible with a non-special k-point kv′. If so, return an αβγ′ value such that kv = kv′(αβγ′).

TODO: This method should eventually be merged with the equivalently named method in PhotonicBandConnectivity/src/connectivity.jl, which handles everything more correctly, but currently has a slightly incompatible API.

Crystalline.isapproxin — Method

isapproxin(x, itr) --> Bool

Determine whether x ∈ itr with approximate equality.

Crystalline.lattice_from_mpbparams — Method

lattice_from_mpbparams(filepath::String)

This will load an input file with path filepath that was previously created by prepare_mpbcalc(!) and return the associated lattice as Julia objects.

Output:

    Rs::DirectBasis,
    flat::ModulatedFourierLattice,
    isoval::Float64,
    epsin::Float64,
    epsout::Float64
    kvecs::Vector{SVector{D, Float64}}

Note that flat does not retain information about orbit groupings, since we flatten the orbits into a single vector in lattice2mpb. This doesn't matter as we typically just want to plot the saved lattice (see plot_lattice_from_mpbparams from compat/pyplot.jl).

Crystalline.matching_lgs — Method

matching_lgs(BRS::BandRepSet)

Finds the matching little groups for each k-point referenced in BRS. This is mainly a a convenience accessor, since e.g. littlegroup(::SpaceGroup, ::KVec) could already give the required little groups. The benefit here is that the resulting operator sorting of the returned little group is identical ISOTROPY's, so we can rely on that later on.

Note that the little groups from ISOTROPY do not include copies of operators that would be identical when transformed to a primitive basis. The operators are, however, still given in a conventional basis.

An error is thrown if a referenced little group cannot be found (currently, this can happen for certain k-points in Φ-Ω, see src/specialrepresentationdomain_kpoints.jl)

Crystalline.orbit — Method

orbit(Ws, x)

Computes the orbit of x under a set of point-group operations Ws, i.e. computes the set {gx | g∈G} where g denotes elements of the group G composed of all operations in Ws (possibly iterated, to ensure full coverage).

At the moment, we only consider point group operations; i.e. there are no nonsymmorphic Ws parts.

It is important that Ws and x are given in the same basis.

[W' = PWP⁻¹ if the basis change is from coordinates r to r' = Pr, corresponding to a new set of basis vectors (x̂')ᵀ=x̂ᵀP; e.g., when going from a direct basis representation to a Cartesian one, the basis change matrix is P = [R₁ R₂ R₃], with Rᵢ inserted as column vectors]

Crystalline.parsefraction — Method

parsefraction(str::AbstractString)

Parse a string str, allowing fraction inputs (e.g. "1/2"), return as Float64.

Crystalline.primitivebasismatrix — Function

primitivebasismatrix(cntr::Char, D::Integer) -> ::Matrix{Float64}

Given a centering type cntr and a dimensionality D, calculates a transformation matrix P from a conventional to a primitive unit cell, using dictionary lookup.

Crystalline.read_sgops_xyzt — Function

read_sgops_xyzt(sgnum::Integer, dim::Integer=3)

Obtains the symmetry operations in xyzt format for a given space group number sgnum by reading from json files; see spacegroup for additional details. Much faster than crawling; generally preferred.

Crystalline.relrand — Method

relrand(lims::NTuple{2,Real}, N=1) --> Vector{Float64}

Computes a random number in the range specified by the two-element tuple lims. The random numbers are sampled from two uniform distributions, namely [lims[1], 1.0] and [1.0, lims[2]], in such a way as to ensure that the sampling is uniform over the joint interval [-1/lims[1], -1.0] ∪ [1.0, lims[2]].

This is useful for ensuring an even sampling of numbers that are either smaller or larger than unity. Eg. for x=relrand((0.2,5.0)), x is equally probable to fall in inv(x)∈[1,5] or x∈[1,5].

Crystalline.rotation_order — Method

rotation_order(W::Matrix{<:Real}) --> Int
rotation_order(op::SymOperation)  --> Int

Determine the integer rotation order of a point group operation, input either as a matrix W or op::SymOperation.

The rotation order of

Proper rotations is positive.

Improper (mirrors, inversion, roto-inversions) is negative.

Crystalline.rotation_order_3d — Method

rotation_order_3d(detW::Real, trW::Real) --> Int

Determine the integer rotation order of a 3D point group operation with a 3×3 matrix representation W (alternatively specified by its determinant detW and its trace trW).

The rotation order of

Proper rotations is positive.

Improper (mirrors, inversion, roto-inversions) is negative.

Crystalline.uniquetol — Method

uniquetol(a; kwargs)

Computes approximate-equality unique with tolerance specifiable via keyword arguments kwargs in O(n²) runtime.

Copied from https://github.com/JuliaLang/julia/issues/19147#issuecomment-256981994

Crystalline.wyckbasis — Method

wyckbasis(BRS::BandRepSet) --> Vector{Vector{Int64}}

Computes the (band representation) basis for bands generated by localized orbitals placed at the Wyckoff positions. Any band representation that can be expanded on this basis with positive integer coefficients correspond to a trivial insulators (i.e. deformable to atomic limit). Conversely, bands that cannot are topological, either fragily (some negative coefficients) or strongly (fractional coefficients).