Bravais types and basis systems

Bravais.jl

Missing docstring.

Missing docstring for crystal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for crystalsystem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for directbasis. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bravaistype. Check Documenter's build log for details.

Missing docstring.

Missing docstring for reciprocalbasis. Check Documenter's build log for details.

Missing docstring.

Missing docstring for primitivize(::Bravais.AbstractBasis, ::Integer). Check Documenter's build log for details.

Missing docstring.

Missing docstring for primitivize(::DirectBasis, ::Char). Check Documenter's build log for details.

Missing docstring.

Missing docstring for primitivize(::ReciprocalBasis, ::Char). Check Documenter's build log for details.

Missing docstring.

Missing docstring for conventionalize(::DirectBasis, ::Char). Check Documenter's build log for details.

Extensions of Bravais.jl methods

Bravais.transform — Function

transform(v::T, P::AbstractMatrix) where T<:AbstractVec --> v′::T

Return a transformed coordinate vector v′ from an original coordinate vector v using a basis change matrix P.

Note that a basis change matrix P transforms direct coordinate vectors (RVec) as $\mathbf{r}' = \mathbf{P}^{-1}\mathbf{r}$ but transforms reciprocal coordinates (KVec) as $\mathbf{k}' = \mathbf{P}^{\mathrm{T}}\mathbf{k}$ (see e.g. ITA7 Secs. 1.5.1.2 and 1.5.2.1).

transform(op::SymOperation, P::Matrix{<:Real}, 
          p::Union{Vector{<:Real}, Nothing}=nothing,
          modw::Bool=true)                          --> SymOperation

Transforms a op $= \{\mathbf{W}|\mathbf{w}\}$ by a rotation matrix P and a translation vector p (can be nothing for zero-translations), producing a new symmetry operation op′ $= \{\mathbf{W}'|\mathbf{w}'\}$ (see ITA6 Sec. 1.5.2.3):

$\{\mathbf{W}'|\mathbf{w}'\} = \{\mathbf{P}|\mathbf{p}\}^{-1}\{\mathbf{W}|\mathbf{w}\} \{\mathbf{P}|\mathbf{p}\}$

with

$\mathbf{W}' = \mathbf{P}^{-1}\mathbf{W}\mathbf{P}$ and $\mathbf{w}' = \mathbf{P}^{-1}(\mathbf{w}+\mathbf{W}\mathbf{p}-\mathbf{p})$

By default, the translation part of op′, i.e. $\mathbf{w}'$, is reduced to the range $[0,1)$, i.e. computed modulo 1. This can be disabled by setting modw = false (default, modw = true).

See also primitivize and conventionalize.

Bravais.primitivize — Method

primitivize(v::T, cntr::Char) where T<:AbstractVec{D} --> v′::T

Transforms a conventional coordinate vector v to a standard primitive basis (specified by the centering type cntr), returning the primitive coordinate vector v′.

Note that a basis change matrix $P$ (as returned e.g. by Crystalline.primitivebasismatrix) transforms direct coordinate vectors (RVec) as $\mathbf{r}' = \mathbf{P}^{-1}\mathbf{r}$ but transforms reciprocal coordinates (KVec) as $\mathbf{k}' = \mathbf{P}^{\mathrm{T}}\mathbf{k}$ (see e.g. ITA7 Sec. 1.5.1.2 and 1.5.2.1). Recall also the distinction between transforming a basis and the coordinates of a vector.

Bravais.conventionalize — Method

conventionalize(v′::T, cntr::Char) where T<:AbstractVec{D} --> v::T

Transforms a primitive coordinate vector v′ back to a standard conventional basis (specified by the centering type cntr), returning the conventional coordinate vector v.

See also primitivize(::AbstractVec, ::Char and transform.