Elementary band representations

Crystalline.jl provides an interface to access the elementary band representations (EBRs) hosted by the Bilbao Crystallographic Server's BANDREP program via bandreps. Please cite the original research (listed in the associated docstrings).

As an example, we can obtain the all inequivalent EBRs in space group 219 (F-43c) with:

using Crystalline

brs = bandreps(219, 3) # space group 219 (dimension 3)

BandRepSet (⋕219): 10 BandReps, sampling 14 LGIrreps (spin-1 w/ TR)
──────┬────────────────────────────────────────────────
      │ 8a   8a   8a  8b   8b   8b  24c  24c  24d  24d
      │ A   ¹E²E  T   A   ¹E²E  T    A    B    A    B
──────┼────────────────────────────────────────────────
 Γ₁   │ 1    ·    ·   1    ·    ·    1    ·    1    ·
 Γ₂   │ 1    ·    ·   1    ·    ·    ·    1    ·    1
 Γ₃   │ ·    2    ·   ·    2    ·    1    1    1    1
 Γ₄   │ ·    ·    1   ·    ·    1    ·    1    ·    1
 Γ₅   │ ·    ·    1   ·    ·    1    1    ·    1    ·
 X₁   │ 1    2    ·   ·    ·    1    ·    1    2    1
 X₂   │ 1    2    ·   ·    ·    1    1    ·    1    2
 X₃   │ ·    ·    1   1    2    ·    2    1    ·    1
 X₄   │ ·    ·    1   1    2    ·    1    2    1    ·
 X₅   │ ·    ·    2   ·    ·    2    1    1    1    1
 L₁L₂ │ 1    ·    1   1    ·    1    1    1    1    1
 L₃L₃ │ ·    1    1   ·    1    1    1    1    1    1
 W₁W₂ │ 1    2    1   ·    ·    2    1    1    2    2
 W₃W₄ │ ·    ·    2   1    2    1    2    2    1    1
──────┼────────────────────────────────────────────────
 μ    │ 2    4    6   2    4    6    6    6    6    6
──────┴────────────────────────────────────────────────
  KVecs (maximal only): Γ, X, L, W

which returns a BandRepSet, which itself is an AbstractVector of BandReps. This allows us to index into brs easily:

brs[1] # obtain the EBR induced by Wyckoff position 8a with irrep A

2-band BandRep (A↑G at 8a):
 [Γ₁+Γ₂, X₁+X₂, L₁L₂, W₁W₂]

By default, bandreps returns the spinless EBRs with time-reversal symmetry. This behavior can be controlled with the keyword arguments spinful (default, false) and timereversal (default, true). By default, only minimal paths are included in the sampling of k-vectors; additional paths can be obtained by setting the keyword argument allpaths = true (default, false).

The distinct topological classes identifiable from symmetry can can be calculated via classification, which uses the Smith normal form's principle factors:

classification(brs)

"Z₁"

Which demonstrates that the symmetry indicator group of spinless particles with time-reversal symmetry in space group 219 is trivial.

Topology and associated bases

The SymmetryBases.jl package provides tools to analyze topology of symmetry vectors and compute associated Hilbert bases.

API

Crystalline.bandreps — Function

bandreps(sgnum::Integer, D::Integer=3; 
         allpaths::Bool=false, spinful::Bool=false, timereversal::Bool=true)

Returns the elementary band representations (EBRs) as a BandRepSet for space group sgnum and dimension D.

Keyword arguments

allpaths: include a minimal sufficient set (false, default) or all (true) k-vectors.

spinful: single- (false, default) or double-valued (true) irreps, as appropriate for spinless and spinful particles, respectively. Only available for D=3.

timereversal: assume presence (true, default) or absence (false) of time-reversal symmetry.

References

3D EBRs are obtained from the Bilbao Crystallographic Server's BANDREP program; please reference the original research papers noted there if used in published work.

Crystalline.classification — Function

classification(BRS_or_F::Union{BandRepSet, Smith}) --> String

Return the symmetry indicator group $X^{\text{BS}}$ of an EBR basis F_or_BRS, provided as a BandRepSet or Smith decomposition.

Technically, the calculation answers the question "what direct product of $\mathbb{Z}_n$ groups is the the quotient group $X^{\text{BS}} = \{\text{BS}\}/\{\text{AI}\}$ isomorphic to?" (see Po, Watanabe, & Vishwanath, Nature Commun. 8, 50 (2017) for more information).

Crystalline.basisdim — Function

basisdim(BRS::BandRepSet) --> Int

Return the dimension of the (linearly independent parts) of a band representation set. This is $d^{\text{bs}} = d^{\text{ai}}$ in the notation of Po, Watanabe, & Vishwanath, Nature Commun. 8, 50 (2017), or equivalently, the rank of matrix(BRS) over the ring of integers. This is the number of linearly independent basis vectors that span the expansions of a band structure viewed as symmetry data.

Crystalline.matching_littlegroups — Function

matching_littlegroups(BRS::BandRepSet)

Finds the matching little groups for each k-point referenced in BRS. This is mainly a a convenience accessor, since e.g. littlegroup(::SpaceGroup, ::KVec) could already give the required little groups. The benefit here is that the resulting operator sorting of the returned little group is identical ISOTROPY's, so we can rely on that later on.

Returns a Vector{<:LittleGroup} with ordering identical to that of the k-point labels in BRS. Note that this is different from the return type of littlegroups which returns an unordered Dict.

Note 1

The little groups from ISOTROPY do not include copies of operators that would be identical when transformed to a primitive basis. The operators are, however, still given in a conventional basis.

Note 2

An error is thrown if a referenced little group cannot be found (currently, this can happen for certain k-points in $Φ-Ω$, see src/specialrepresentationdomain_kpoints.jl)