Bravais types and basis systems

API

Crystalline.primitivebasismatrix — Method

primitivebasismatrix(cntr::Char, ::Val{D}=Val(3)) -> SMatrix{D,D,Float64}

Return the transformation matrix P that transforms a conventional unit cell with centering cntr to the corresponding primitive unit cell (in dimension D).

Crystalline.relrand — Method

relrand(lims::NTuple{2,Real}) --> Float64

Computes a random number in the range specified by the two-element tuple lims. The random numbers are sampled from two uniform distributions, namely [lims[1], 1] and [1, lims[2]], in such a way as to ensure that the sampling is uniform over the joint interval [-1/lims[1], -1] ∪ [1, lims[2]].

This is useful for ensuring an even sampling of numbers that are either smaller or larger than unity. Eg. for x = relrand((0.2,5.0)), x is equally probable to fall in inv(x)∈[1,5] or x∈[1,5].