Bravais types and basis systems

Crystalline.crystal — Function

crystal(a,b,c,α,β,γ) --> Rs::DirectBasis{3}

Calculate basis vectors $\mathbf{R}_1$, $\mathbf{R}_2$, $\mathbf{R}_3$ in a 3D Cartesian basis for a right-handed coordinate system with specified basis vector lengths a, b, c (associated with $\mathbf{R}_1$, $\mathbf{R}_2$, & $\mathbf{R}_3$, respectively) and specified interaxial angles α $= ∠(\mathbf{R}_2,\mathbf{R}_3)$, β $= ∠(\mathbf{R}_3,\mathbf{R}_1)$, γ $= ∠(\mathbf{R}_1,\mathbf{R}_2)$, with $∠$ denoting the angle between two vectors.

For definiteness, the $\mathbf{R}_1$ basis vector is oriented along the $x$-axis of the Cartesian coordinate system, and the $\mathbf{R}_2$ axis is placed in the $xy$-plane.

crystal(a,b,γ) --> DirectBasis{2}

Calculate basis vectors $\mathbf{R}_1$, $\mathbf{R}_2$ in a 2D Cartesian basis for a right-handed coordinate system with specified basis vector lengths a, b (associated with $\mathbf{R}_1$ & $\mathbf{R}_2$, respectively) and specified interaxial angle γ $= ∠(\mathbf{R}_1,\mathbf{R}_2)$.

For definiteness, the $\mathbf{R}_1$ basis vector is oriented along the $x$-axis of the Cartesian coordinate system.

crystal(a)  --> DirectBasis{1}

Return a one-dimensional crystal with lattice period a.

Crystalline.crystalsystem — Function

crystalsystem(R::DirectBasis{D})

Determine the crystal system of a point lattice specified in a conventional DirectBasis{D} of dimension D using Tables 2.1.2.1, 9.1.7.1, & 9.1.7.2 of the International Tables of Crystallography, Volume 1 (ITA). There are 4 crystal systems in 2D and 7 in 3D (see ITA 2.1.2(iii)):

D	System	Conditions	Free parameters

1D	linear	none	a

2D	square	a=b & γ=90°	a

rectangular	γ=90°	a,b

hexagonal	a=b & γ=120°	a

oblique	none	a,b,γ

3D	cubic	a=b=c & α=β=γ=90°	a

hexagonal	a=b & α=β=90° & γ=120°	a,c

trigonal	a=b & α=β=90° & γ=120°	a,c (a,α for hR)

tetragonal	a=b & α=β=γ=90°	a,c

orthorhombic	α=β=γ=90°	a,b,c

monoclinic	α=γ=90°	a,b,c,β≥90°

triclinic	none	a,b,c,α,β,γ

The DirectBasis input is assumed to use conventional basis vectors; i.e. not necessarily primitive. For primitive basis vectors, the crystal system can be further reduced into 5 Bravais types in 2D and 14 in 3D.

Crystalline.directbasis — Function

directbasis(sgnum, D=3;    abclims, αβγlims)
directbasis(sgnum, Val(D); abclims, αβγlims) --> DirectBasis{D}

Return a (conventional) DirectBasis for a crystal compatible with the space group number sgnum and dimensionality D. Free parameters in the lattice vectors are chosen randomly, with limits optionally supplied in abclims (lengths) and αβγlims (angles). By convention, the length of the first lattice vector (a) is set to unity, such that the second and third (b and c) lattice vectors' lengths are relative to the first.

Limits on the relative uniform distribution of lengths b and c can be specified as 2-tuple kwarg abclims; similarly, limits on the angles α, β, γ can be set via αβγlims (only affects oblique, monoclinic, & triclinic lattices).

Crystalline.bravaistype — Function

bravaistype(sgnum::Integer) -> String
bravaistype(sgnum::Integer, D::Integer; normalize) -> String

Return the Bravais type of sgnum in dimension D as a string (as the concatenation of the single-character crystal abbreviation and the centering type).

Keyword arguments

If the centering type associated with sgnum is 'A', we can choose (depending on the keyword argument normalize, defaulting to false) to "normalize" to the centering type 'C', since the difference between A and C centering only amounts to a basis change. With normalize=true we then have only the canonical 14 Bravais type, i.e. unique(bravaistype.(1:230, 3), normalize=true) returns only 14 distinct types, rather than 15.

This only affects space groups 38-41 (normalizing their conventional Bravais types from oA to oC).

Crystalline.reciprocalbasis — Function

reciprocalbasis(Rs::DirectBasis{D}) --> Gs::ReciprocalBasis{D}

Calculates the reciprocal basis associated with a DirectBasis Rs (alternatively supplied as an NTuple of Vectors). Returns a ReciprocalBasis.

Crystalline.primitivize — Method

primitivize(Vs::Basis, sgnum::Integer) --> Rs′::Basis

Transforms a conventional Basis (either DirectBasis or ReciprocalBasis) Vs into its primitive equivalent Vs′, provided that its centering differs from the conventional (P or p), by inferring the Bravais type from the space group number sgnum and applying an applying an appropriate (Basis-type specific) transformation.

Crystalline.primitivize — Method

primitivize(Rs::DirectBasis, cntr::Char) --> Rs′::DirectBasis

Transforms a conventional DirectBasis Rs into its primitive equivalent Rs′, with the transformation dependent on the centering type cntr (P, I, F, R, A, C, and p, c); for centering P and p, the conventional and primive bases coincide.

Crystalline.primitivize — Method

primitivize(Gs::ReciprocalBasis, cntr::Char) --> Gs′::ReciprocalBasis

Calculates the primitive reciprocal basis associated with an assumed conventional reciprocal basis Gs::ReciprocalBasis with centering type cntr.

Crystalline.conventionalize — Method

conventionalize(Rs′::DirectBasis, cntr::Char) --> Rs::DirectBasis

Transforms a primitive DirectBasis Rs′ into its conventional equivalent Rs, with the transformation dependent on the centering type cntr (P, I, F, R, A, C, and p, c); for centering P and p, the conventional and primive bases coincide.

Crystalline.transform — Function

transform(op::SymOperation, P::Matrix{<:Real}, 
          p::Union{Vector{<:Real}, Nothing}=nothing,
          modw::Bool=true)                          --> SymOperation

Transforms a op $= \{\mathbf{W}|\mathbf{w}\}$ by a rotation matrix P and a translation vector p (can be nothing for zero-translations), producing a new symmetry operation op′ $= \{\mathbf{W}'|\mathbf{w}'\}$ (see ITA6 Sec. 1.5.2.3):

$\{\mathbf{W}'|\mathbf{w}'\} = \{\mathbf{P}|\mathbf{p}\}^{-1}\{\mathbf{W}|\mathbf{w}\} \{\mathbf{P}|\mathbf{p}\}$

with

$\mathbf{W}' = \mathbf{P}^{-1}\mathbf{W}\mathbf{P}$ and $\mathbf{w}' = \mathbf{P}^{-1}(\mathbf{w}+\mathbf{W}\mathbf{p}-\mathbf{p})$

By default, the translation part of op′, i.e. $\mathbf{w}'$, is reduced to the range $[0,1)$, i.e. computed modulo 1. This can be disabled by setting modw = false (default, modw = true).

See also primitivize and conventionalize.

transform(v::T, P::AbstractMatrix) where T<:AbstractVec --> v′::T

Return a transformed coordinate vector v′ from an original coordinate vector v using a basis change matrix P.

Note that a basis change matrix P transforms direct coordinate vectors (RVec) as $\mathbf{r}' = \mathbf{P}^{-1}\mathbf{r}$ but transforms reciprocal coordinates (KVec) as $\mathbf{k}' = \mathbf{P}^{\mathrm{T}}\mathbf{k}$ (see e.g. ITA7 Secs. 1.5.1.2 and 1.5.2.1).

Crystalline.primitivize — Method

primitivize(v::T, cntr::Char) where T<:AbstractVec{D} --> v′::T

Transforms a conventional coordinate vector v to a standard primitive basis (specified by the centering type cntr), returning the primitive coordinate vector v′.

Note that a basis change matrix $P$ (as returned e.g. by Crystalline.primitivebasismatrix) transforms direct coordinate vectors (RVec) as $\mathbf{r}' = \mathbf{P}^{-1}\mathbf{r}$ but transforms reciprocal coordinates (KVec) as $\mathbf{k}' = \mathbf{P}^{\mathrm{T}}\mathbf{k}$ (see e.g. ITA7 Sec. 1.5.1.2 and 1.5.2.1). Recall also the distinction between transforming a basis and the coordinates of a vector.

Crystalline.conventionalize — Method

conventionalize(v′::T, cntr::Char) where T<:AbstractVec{D} --> v::T

Transforms a primitive coordinate vector v′ back to a standard conventional basis (specified by the centering type cntr), returning the conventional coordinate vector v.

See also primitivize(::AbstractVec, ::Char and transform.